gmlgo: update gmlgo and worm to utilize fixed delta time
Fixes #93

+++ examples/worm/game/wall_spawner.go
@@ -85,6 +85,6 @@ func (self *WallSpawner) Update(roomInstanceIndex gml.RoomInstanceIndex) {
 		//    additional_time += room_speed * 4;
 		//    flight_extended_timer_used = true;
 		//}
-		self.SpawnWallTimer.Set(ticksTillNextWall)
+		self.SpawnWallTimer.Set(float64(ticksTillNextWall))
 	}
 }
